+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Breaking Changes
   * Drop support for ruby 1.9.3
   * HTTP Basic Auth is now disabled by default (use allow_http_basic_auth to enable)
+  * 'httponly' and 'secure' cookie options are enabled by default now
 
 # Fixed
   * ensure that login field validation uses correct locale (@sskirby)